Split the tasks and DRY the gulpfile
