+++ development/index.html
@@ -1,0 +1,69 @@
+<!doctype html>
+<html class="no-js" lang="">
+  <head>
+    <meta charset="utf-8">
+    <meta http-equiv="x-ua-compatible" content="ie=edge">
+    <title></title>
+    <meta name="description" content="">
+    <meta name="viewport" content="width=device-width,initial-scale=1">
+
+    <link rel="apple-touch-icon" href="apple-touch-icon.png">
+    <!-- Place favicon.ico in the root directory -->
+
+    <link rel="stylesheet" href="css/style.css">
+    <script src="js/vendor/modernizr-2.8.3.min.js"></script>
+
+  </head>
+  <body class="global-wrapper">
+    <header class="main-header">
+      <div class="logo">
+        <a href="index.html">
+          Proj
+        </a>
+      </div>
+      <nav class="main-nav">
+        <ul>
+          <li><a href="#">home</a></li>
+          <li><a href="#">page</a></li>
+          <li><a href="#">page</a></li>
+          <li><a href="#">page</a></li>
+        </ul>
+      </nav>
+    </header>
+    <main class="main-content">
+      <section class="info-box">
+        
+      </section>
+      <section class="article-container">
+
+        <article>
+        <h1 class="content-title">
+          Content components
+
+        </h1></article>
+
+      </section>
+    </main>
+    <footer class="main-footer">
+
+    </footer>
+
+    <!-- Js plugins -->
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/2.0.0/handlebars.js"></script>
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js"></script>
+    <script>window.jQuery || document.write('<script src="js/vendor/jquery-1.11.3.min.js"><\/script>')</script>
+    <script async src="js/main.js" onload="myInit()"></script>
+    <script src="js/plugins.js"></script>  
+    <script src="js/ui/headings.ui.js"></script>
+
+    <!-- Google Analytics: change UA-XXXXX-X to be your site's ID. -->
+    <script>
+        (function(b,o,i,l,e,r){b.GoogleAnalyticsObject=l;b[l]||(b[l]=
+        function(){(b[l].q=b[l].q||[]).push(arguments)});b[l].l=+new Date;
+        e=o.createElement(i);r=o.getElementsByTagName(i)[0];
+        e.src='https://www.google-analytics.com/analytics.js';
+        r.parentNode.insertBefore(e,r)}(window,document,'script','ga'));
+        ga('create','UA-XXXXX-X','auto');ga('send','pageview');
+    </script>
+  </body>
+</html>
